Extract aya-build for building eBPF crates
We'll use this in the template and book to avoid duplicating all the
code.
