+++ aya-build/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "aya-build"
+version = "0.1.0"
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+edition.workspace = true
+
+[dependencies]
+anyhow = { workspace = true }
+cargo_metadata = { workspace = true }
